refactor: NotificationManager 분리
